Move Document from cats to core
This is a first cut at allowing non-Cats code to use the Document type class.
