+++ core/src/main/scala/org/typelevel/paiges/Document.scala
@@ -1,8 +1,5 @@
 package org.typelevel.paiges
 
-import cats.Cartesian
-import cats.functor.Contravariant
-
 trait Document[A] { self =>
   def document(a: A): Doc
 
@@ -52,17 +49,4 @@ object Document {
   object FromToString extends Document[Any] {
     def document(any: Any): Doc = Doc.text(any.toString)
   }
-
-  implicit val contravariantDocument: Contravariant[Document] =
-    new Contravariant[Document] {
-      def contramap[A, Z](d: Document[A])(f: Z => A): Document[Z] = d.contramap(f)
-    }
-
-  def cartesianDocument(sep: Doc): Cartesian[Document] =
-    new Cartesian[Document] {
-      def product[A, B](fa: Document[A], fb: Document[B]): Document[(A, B)] =
-        Document.instance { case (a, b) =>
-          fa.document(a) + sep + fb.document(b)
-        }
-    }
 }
