+++ core/src/test/scala/org/typelevel/paiges/DocumentTests.scala
